+++ include/parser.h
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/24 12:19:42 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/25 22:58:05 by tsishika         ###   ########.fr       */
+/*   Updated: 2023/09/29 11:21:25 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,33 +14,10 @@
 # define PARSER_H
 
 # include "libft.h"
-# include "lexer.h"
-
-typedef enum e_node_type
-{
-	ND_CMD,
-	ND_PIPE
-}	t_node_type;
-
-typedef struct s_redirect
-{
-	char			*file;
-	t_token_type	type;
-}	t_redirect;
-
-typedef struct s_ast	t_ast;
-struct s_ast
-{
-	t_node_type	type;
-	t_token		*argv;
-	t_list		*redir_lst;
-	t_ast		*left;
-	t_ast		*right;
-	t_token		*lst;
-};
+# include "type.h"
 
 t_ast	*ast_new_node(t_node_type type, t_ast *left, t_ast *right);
 t_ast	*ast_new_node_cmd(t_token *lst);
-t_ast	*parse_token(t_token *lst);
+t_ast	*parse_token(t_context *ctx);
 
 #endif
